+++ netdiag/pages/__init__.py
@@ -4,3 +4,4 @@ from .netinfo import NetInfoPage
 from .arp import ArpPage
 from .log import LogPage
 from .analyze import AnalyzePage
+from .portscan import PortScanPage